feat: move countly static file to utils folder
