+++ src/components/forms/feedback-survey/option-b/Form.vue
@@ -72,7 +72,7 @@
 </template>
 <script>
 import translations from '../../../../static/translations'
-import { EVENTS } from '../../../../static/countly'
+import countly from '../../../../utils/countly'
 import { getTutorialByUrl } from '../../../../utils/tutorials'
 import settings from '../../../../utils/settings'
 import Question from '../Question.vue'
@@ -123,24 +123,21 @@ export default {
       }
     },
     onDismiss: function () {
-      this.trackEvent(EVENTS.TUTORIAL_FEEDBACK_SURVEY_DISMISSED, {
+      this.trackEvent(countly.events.TUTORIAL_FEEDBACK_SURVEY_DISMISSED, {
         numberOfQuestionsAnswered: this.currentStep === -1 ? 0 : this.currentStep,
         surveyCompleted: this.currentStep === this.maximumStep
       })
 
       this.dismissed = true
     },
-    trackEvent: function (event, opts = {}) {
+    trackEvent: function (event, data) {
       const tutorialFeedbackSurveyOption = settings.abTesting.get(settings.abTesting.TUTORIAL_FEEDBACK_SURVEY)
 
-      window.Countly.q.push(['add_event', {
-        key: event,
-        segmentation: {
-          path: this.$route.path,
-          option: tutorialFeedbackSurveyOption,
-          ...opts
-        }
-      }])
+      countly.trackEvent(event, {
+        path: this.$route.path,
+        option: tutorialFeedbackSurveyOption,
+        ...data
+      })
     }
   }
 }
