+++ src/utils/countly.js
@@ -1,4 +1,4 @@
-export const EVENTS = {
+export const events = {
   CODE_RESET: 'resetCode',
   CODE_SUBMIT_WRONG: 'submitWrongCode',
   CHOICE_SUBMIT_WRONG: 'submitWrongChoice',
@@ -10,4 +10,16 @@ export const EVENTS = {
   TUTORIAL_FEEDBACK_SURVEY_COMPLETED: 'tutorialFeedbackSurveyCompleted',
   TUTORIAL_FEEDBACK_SURVEY_DISMISSED: 'tutorialFeedbackSurveyDismissed',
   TUTORIAL_FEEDBACK_SURVEY_AB_TESTING: 'tutorialFeedbackSurveyABTesting'
+}
+
+export function trackEvent (event, data = {}) {
+  window.Countly.q.push(['add_event', {
+    key: event,
+    segmentation: data
+  }])
+}
+
+export default {
+  events,
+  trackEvent
 }
